fixed kill 8080

+++ codedeploy/beforeInstall.sh
@@ -1,5 +1,11 @@
 #!/bin/bash -v
 
 # shellcheck disable=SC2046
-sudo kill -9 $(lsof -t -i:8080)
+if [ -z $(lsof -t -i:8080) ]
+then
+      echo "nothing runs on 8080"
+else
+      sudo kill -9 $(lsof -t -i:8080)
+fi
+
 source /env/properties.sh
